+++ py/frequenz/api/platformassets/__init__.py
@@ -1,0 +1,4 @@
+# License: MIT
+# Copyright © 2025 Frequenz Energy-as-a-Service GmbH
+
+"""Frequenz gRPC API to retrieve platform assets information."""