+++ py/frequenz/api/platformassets/v1alpha1/__init__.py
@@ -1,0 +1,4 @@
+# License: MIT
+# Copyright © 2025 Frequenz Energy-as-a-Service GmbH
+
+"""Frequenz Platform Assets API v1alpha1."""